Brand new images and clothing

+++ app/src/main/java/no/ntnu/prisonesc/PhysicsObject.java
@@ -203,7 +203,7 @@ public class PhysicsObject {
         this.posY += height;
     }
 
-    public void addGliderFactor(float amount) {
+    public void addGliderFactor(double amount) {
         this.gliderFactor += amount;
     }
